requirements.txt
Flask
spotipy
qrcode
reportlab
python-dotenv
WeasyPrint
tailwindcss
qrcode
reportlab
flask-cors
flask-socketio
reportlab
Pillow